dengue-chat-bot-LangChainKB/requirements.txt
langchain
PyPDF2
faiss-cpu
tiktoken
langchain-community
langchain-google-genai
python-dotenv
google-generativeai
flask
flask-cors
requests